Add hafnium-lts-automation
This job is triggered when a new LTS branch `lts-vX.Y` is created in the
Hafnium repository. It automatically creates the corresponding LTS branch
(`lts-vX.Y`) in the following submodules using the tag `vX.Y`:

  - hafnium/project/reference
  - hafnium/prebuilts
  - hafnium/third_party/googletest
  - hafnium/third_party/dtc
  - ci/hafnium-ci-scripts
  - ci/hafnium-job-configs

If the new LTS branch does not exist in the TF-A repository, the job will
also trigger the `tf-a-lts-automation` job to create it.

Sandbox runs are also supported. When a branch `sandbox/lts-vX.Y` is
created, the job will create the corresponding branch in the submodules
using the tag `sandbox/vX.Y`.

Change-Id: I7b379bbda3a4213a5912bb67e2ac8b98b693b1bf
Signed-off-by: Arthur She <arthur.she@linaro.org>